Update Console.c
By adding a if statement in function 'void addHistory(char *command)' blank commands will not be added to the history of the commands.
By shifting the statement 'currentCommandId--;' from line 414 to line 422, when up key is pressed, most recent command is shown. Otherwise it was showing the second most recent command when up key was pressed.